requirements.txt
dask
matplotlib==3.7.1
matplotlib-inline==0.1.7
matplotlib-venn==1.1.1
numpy==1.26.4
pandas==2.1.4
pandas-datareader==0.10.0
pandas-gbq==0.23.1
pandas-stubs==2.1.4.231227
scikit-image==0.23.2
scikit-learn==1.3.2
scipy==1.13.1
seaborn==0.13.1
sklearn-pandas==2.2.0
imbalanced-learn==0.12.3
